[front] move activateCreditPricedFreePlan to subscription lib
